updated js toasts and the customer website (now waits for NFC cards)

+++ transactify_service/run_on_host.sh
@@ -9,13 +9,18 @@ export DJANGO_SUPERUSER_PASSWORD="admin"
 export DJANGO_SUPERUSER_USERNAME="admin"
 export DJANGO_SUPERUSER_EMAIL="$DJANGO_SUPERUSER_PASSWORD@$DJANGO_SUPERUSER_USERNAME.com"
 
-export "RUN_SERVER"="false"
-psql -h $PGHOST -U $PGUSER -p 5432 -d 'postgres' -c "DROP DATABASE IF EXISTS \"$MAIN_DB\";"
-psql -h $PGHOST  -U $PGUSER -p 5432 -d 'postgres' -c "CREATE DATABASE \"$MAIN_DB\" OWNER \"$PGUSER\";"
-
-# Run migration
-python manage.py makemigrations && python manage.py migrate
-python manage.py createsuperuser --noinput
-export "RUN_SERVER"="true"
 export  "INIT_DATA"=1
+if [ "$INIT_DATA" = "1" ]; then
+    echo "Initializing data..."
+    export "RUN_SERVER"="false"
+    psql -h $PGHOST -U $PGUSER -p 5432 -d 'postgres' -c "DROP DATABASE IF EXISTS \"$MAIN_DB\";"
+    psql -h $PGHOST  -U $PGUSER -p 5432 -d 'postgres' -c "CREATE DATABASE \"$MAIN_DB\" OWNER \"$PGUSER\";"
+
+    # Run migration
+    python manage.py makemigrations && python manage.py migrate
+    python manage.py createsuperuser --noinput
+fi
+
+export "RUN_SERVER"="true"
+
 daphne -b 0.0.0.0 -p 8880 transactify_service.asgi:application
